pyproject.toml
[tool.poetry]
name = "loh-lenhs"
version = "1.0.0"
description = "Software de análises para o Instituto de Pesquisa Hidráulicas da UFRGS."
authors = ["Leandro Monteiro <neofahrenheituk@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
wxpython = "^4.2.3"
openpyxl = "^3.1.5"
numpy = "^2.3.2"
scipy = "^1.16.1"
matplotlib = "^3.10.5"
intersect = "^1.3.2"
unidecode = "^1.4.0"
fpdf2 = "^2.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
app = "app:main"